Fixed warning
git-svn-id: https://wush.net/svn/mdaus-oss/branches/EXP-20100309@79 9d772353-a6c9-4ee9-b2cc-df47a7649cd5

+++ modules/c++/samples/test_parse_xml.cpp
@@ -318,7 +318,7 @@ void generateKMLForSICD(xml::lite::Element* docXML, six::sicd::ComplexData* data
     // Time associated with start
     six::DateTime dateTime = data->timeline->collectStart;
 
-    int durationInSeconds = data->timeline->collectDuration;
+    int durationInSeconds = (int)data->timeline->collectDuration;
     six::Vector3 atTimeX;
     for (unsigned int i = 0; i < durationInSeconds; ++i)
     {
